requirements.txt
Flask==0.12.2
Flask-JWT==0.3.2
flask-swagger-ui
kubernetes
python-etcd
python-jenkins
prometheus_client
pyyaml
requests
gunicorn

# heat provisioner
python-keystoneclient
python-heatclient

# dev
coveralls
faker
flake8
flask-shell-ipython
ipython
pytest
pytest-cov
pytest-env
pytest-flask
pytest-ordering
sphinx
sphinx-autobuild
sphinx_rtd_theme

#GCE
google-api-python-client==1.6.4
google-auth==1.2.1
google-auth-httplib2==0.0.3

#AKS
azure==2.0.0
azure-mgmt-containerservice
